# Conflicts:
#	api-gateway/pom.xml
#	authentication-service/src/main/java/com/example/authentication/config/SecurityConfig.java
#	authentication-service/src/test/java/com/example/authentication/integration/controller/AuthenticationControllerTest.java
#	authentication-service/src/test/java/com/example/authentication/integration/mocks/ProfileMock.java
#	tweet-service/src/main/java/com/example/controller/TweetController.java
#	tweet-service/src/main/java/com/example/service/TweetService.java

+++ tweet-service/src/main/java/com/example/dto/response/TweetResponse.java
@@ -16,5 +16,5 @@ public class TweetResponse {
     private String text;
     private String username;
     private LocalDateTime creationDate;
-    private Long likes;
+    private Integer likes;
 }
